requirements.txt
# FastAPI and related dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database dependencies
motor==3.3.2
beanie==1.23.6
pymongo==4.6.0

# Google AI dependencies
google-generativeai==0.8.5

# Environment and configuration
python-dotenv==1.0.0

# Validation and serialization
pydantic==2.5.0

# HTTP client for additional API calls if needed
httpx==0.25.2

# Logging and utilities
loguru==0.7.2

# Streamlit for UI
streamlit==1.28.1
pandas==2.1.3
plotly==5.17.0

# Development dependencies (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.11.0
# flake8==6.1.0 